Breakdown parser into smaller functions
parser.go used to be one large file with multiple functions like `filter`, `get`
, `select` etc. It has been broken down into smaller functions for easier identification and development
