+++ common/lexer_test.go
