+++ common/matchers.go
@@ -12,18 +12,10 @@ const (
 	countRegex            = `^#$`
 )
 
-type MatcherExp struct {
-	KeyReg       *regexp.Regexp
-	FilterReg    *regexp.Regexp
-	SelectionReg *regexp.Regexp
-	CountReg     *regexp.Regexp
-	SliceReg     *regexp.Regexp
-}
-
-var Matcher = &MatcherExp{
-	KeyReg:       regexp.MustCompile(keyRegex),
-	FilterReg:    regexp.MustCompile(filterRegex),
-	SelectionReg: regexp.MustCompile(selectionRegex),
-	CountReg:     regexp.MustCompile(countRegex),
-	SliceReg:     regexp.MustCompile(sliceRegex),
-}
+var (
+	KeyReg       = regexp.MustCompile(keyRegex)
+	FilterReg    = regexp.MustCompile(filterRegex)
+	SelectionReg = regexp.MustCompile(selectionRegex)
+	CountReg     = regexp.MustCompile(countRegex)
+	SliceReg     = regexp.MustCompile(sliceRegex)
+)
